MoveMage
moved carly's mage code out because it needs work and because we may not use mages with current strategy.
